+++ HannaogHelga/rammeverk/source/control.h
@@ -60,13 +60,13 @@ int get_next_floor(int current_floor, elev_motor_direction_t direction);
 * @param target_floor The next floor the elevator is going to.
 * @return next direction for the elevator (up/down).
 */
-elev_motor_direction_t get_next_direction(int current_floor, int target_floor);
+elev_motor_direction_t get_next_direction(double current_floor, int target_floor);
 
 /**
 * @brief Checks if the elevator should stop at the current floor.
 * @param current_floor The floor the elevator is located at.
 * @param direction The direction the elevator has (up/down).
-*return 1 if it should stop at the floor, 0 if not. 
+*return 1 if it should stop at the floor, 0 if not.
 */
 int should_I_stop(int current_floor, elev_motor_direction_t direction);
